[GraphQL/Package Resolver] Early check for type param arity mismatch
Move the test for type parameter arity mismatch into the first phase
of layout calculation (while fetching packages). This helps avoid
wasting work if we know the resolution will not succeed, and it also
makes the likelihood of exceeding the type parameter width limit
exceedingly low in practice, because all parameter widths are checked
against the arity of a type that is actually loaded (so it would
require the RPC's limit to be set lower than the limit on-chain).

Test Plan:

This is a refactor of existing functionality so relies on existing
tests:

```
sui-package-resolver$ cargo nextest run
```